+++ CHANGELOG.md
@@ -5,6 +5,17 @@ Based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), following [Se
 
 ## [Unreleased]
 
+## [0.1.3]
+- fix plugin layout - import of fixtures
+- fix the gatling fixture
+- add more helper methods
+  - wait for job / deployment
+  - api proxy access
+
+## [0.1.2] 2020-07-07
+- no code changes
+- tools include pre-commit
+
 ## [0.1.1] 2020-07-06
 - First release version